treewide: replace pythonPath for non-buildPython* packages
Replace pythonPath with requiredPythonModules for non-buildPython*
packages.

In this case, it is just a name of the attribute, and it does not matter
what it is because `wrapPythonProgramsIn` is invoked explicitly.

In a future commit wrapPython hook will be replaced by one that does not
require an explicit invocatoin of `wrapPythonProgramsIn`.